+++ Echo/src/utopia/echo/controller/client/OllamaClient.scala
@@ -7,8 +7,8 @@ import utopia.disciple.model.error.RequestFailedException
 import utopia.disciple.model.request.Timeout
 import utopia.echo.controller.Chat
 import utopia.echo.model.llm.LlmDesignator
-import utopia.echo.model.request.generate.Prompt
-import utopia.echo.model.request.llm.{CreateModelRequest, ListModelsRequest, ShowModelRequest}
+import utopia.echo.model.request.ollama.generate.Prompt
+import utopia.echo.model.request.ollama.llm.{CreateModelRequest, ListModelsRequest, ShowModelRequest}
 import utopia.echo.model.response.ollama.llm.StreamedStatus
 import utopia.flow.async.TryFuture
 import utopia.flow.collection.immutable.Empty