+++ Echo/src/utopia/echo/model/response/ollama/ResponseStatistics.scala
@@ -1,6 +1,6 @@
 package utopia.echo.model.response.ollama
 
-import utopia.echo.model.request.generate.GenerateBufferedOrStreamed
+import utopia.echo.model.request.ollama.generate.GenerateBufferedOrStreamed
 import utopia.flow.generic.model.immutable.{Model, Value}
 
 object ResponseStatistics